Move TableLayoutPanelCellPositionTypeConverter to a separate file
